+++ ogc-example/src/main/scala/geotrellis/server/ogc/Main.scala
@@ -1,9 +1,5 @@
 package geotrellis.server.ogc
 
-import geotrellis.proj4.LatLng
-import geotrellis.raster._
-import geotrellis.layer._
-import geotrellis.vector.Extent
 import geotrellis.server.ogc.conf._
 import geotrellis.server.ogc.wms._
 import geotrellis.server.ogc.wcs._
@@ -11,20 +7,15 @@ import geotrellis.server.ogc.wmts._
 
 import cats.effect._
 import cats.implicits._
-
 import com.monovore.decline._
-
 import fs2._
-
 import org.http4s._
 import org.http4s.server._
 import org.http4s.server.blaze.BlazeServerBuilder
 import org.http4s.server.middleware.{CORS, CORSConfig}
 import org.http4s.syntax.kleisli._
-
 import pureconfig._
 import pureconfig.generic.auto._
-
 import org.backuity.ansi.AnsiFormatter.FormattedHelper
 
 import scala.concurrent.duration._