+++ ogc/src/main/scala/geotrellis/server/ogc/params/CRSUtils.scala
@@ -1,20 +1,20 @@
 package geotrellis.server.ogc.params
 
+import geotrellis.proj4._
 import cats.data.{Validated, ValidatedNel}
 import Validated._
-import geotrellis.proj4._
 
 import scala.util.Try
 
-object CrsUtils {
+object CRSUtils {
   /** Converts an OGC URN string representing a CRS into a [[geotrellis.proj4.CRS]] instance.
-    * TODO: Move this into geotrellis.proj4 (or future geotrellis.crs)
-    */
+   * TODO: Move this into geotrellis.proj4 (or future geotrellis.crs)
+   */
   def ogcToCRS(crsDesc: String): ValidatedNel[ParamError, CRS] = {
     val code = crsDesc.trim.toLowerCase
     if(code == "wgs84(dd)") {
       Valid(LatLng).toValidatedNel
-    } else if(code.startsWith("urn:ogc:def:crs:EPSG::")) {
+    } else if(code.startsWith("urn:ogc:def:crs:epsg::")) {
       Try(
         CRS.fromEpsgCode(code.split("::")(1).toInt)
       ).toOption match {